requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
# Uncomment below if using SQLite with SQLAlchemy or async
sqlalchemy
databases[sqlite]
# APScheduler (for background jobs, optional)
# apscheduler 